Rename the project to maggid
A maggid is an itinerant preacher — one who tells, as against one who
rules on the law. It suits a server whose only job is to say what
happened, and it beats a name that just lists the technologies.

Package, console scripts, config directory, and launchd label all move.
The config directory move needs the real ~/.config directory renamed
alongside it, or the voice refs and assignments go missing.
