+++ launchd/com.maggid.plist
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  LaunchAgent for the shared TTS daemon.
+  LaunchAgent for the shared maggid daemon.
 
   Install (substitutes the absolute path, which launchd requires):
 
-    sed "s|__VENV__|$PWD/.venv|g" launchd/com.tts-mcp-server.plist \
-      > ~/Library/LaunchAgents/com.tts-mcp-server.plist
-    launchctl load ~/Library/LaunchAgents/com.tts-mcp-server.plist
+    sed "s|__VENV__|$PWD/.venv|g" launchd/com.maggid.plist \
+      > ~/Library/LaunchAgents/com.maggid.plist
+    launchctl load ~/Library/LaunchAgents/com.maggid.plist
 
   This must be a LaunchAgent, not a LaunchDaemon. A LaunchDaemon runs outside
   the user's GUI session and cannot reach the audio device.
@@ -16,11 +16,11 @@
 <plist version="1.0">
 <dict>
     <key>Label</key>
-    <string>com.tts-mcp-server</string>
+    <string>com.maggid</string>
 
     <key>ProgramArguments</key>
     <array>
-        <string>__VENV__/bin/tts-mcp-server</string>
+        <string>__VENV__/bin/maggid</string>
         <string>--transport</string>
         <string>http</string>
         <string>--host</string>
@@ -43,9 +43,9 @@
     <integer>10</integer>
 
     <key>StandardOutPath</key>
-    <string>/tmp/tts-mcp-server.log</string>
+    <string>/tmp/maggid.log</string>
     <key>StandardErrorPath</key>
-    <string>/tmp/tts-mcp-server.err</string>
+    <string>/tmp/maggid.err</string>
 
     <key>ProcessType</key>
     <string>Interactive</string>
