+++ maggid/__init__.py
@@ -1,4 +1,7 @@
-"""TTS MCP server for Claude Code.
+"""Maggid: a TTS MCP server for Claude Code.
+
+A maggid is an itinerant preacher — one who tells, rather than one who rules on
+the law. The name fits a server whose whole job is to say what happened.
 
 Speaks notifications and narration through Chatterbox Turbo on Apple Silicon,
 via MLX-audio. Provides `notify` for short alerts, `speak` for longer narration,
@@ -6,6 +9,6 @@ and `interrupt` to stop playback and discard the backlog. Runs as a per-session
 stdio server or as one shared daemon.
 """
 
-from tts_server.server import init, main
+from maggid.server import init, main
 
 __all__ = ["init", "main"]
